oblivion/pyproject.toml
[project]
name = "oblivion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "fastapi-swagger-dark>=0.0.9",
    "google-genai>=1.73.1",
    "motor>=3.7.1",
    "pydantic>=2.13.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
]

[[tool.mypy.overrides]]
module = "fastapi_swagger_dark.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fastapi_swagger_dark"
ignore_missing_imports = true
